requirements.txt
asgiref==3.8.1
Authlib==1.3.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.5
Django==5.0.4
djangorestframework==3.15.1
exceptiongroup==1.2.0
idna==3.6
iniconfig==2.0.0
josepy==1.14.0
mozilla-django-oidc==4.0.1
packaging==24.0
pluggy==1.4.0
pycparser==2.22
pyOpenSSL==24.1.0
pytest==8.1.1
pytest-django==4.8.0
pytest-mock==3.14.0
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.10.0
urllib3==2.2.1
